# backend/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Go + Gin + Vue 的企业级任务协作系统后端API\u003cbr/\u003e\n功能模块: 项目管理 | 任务看板 | 甘特图 | 日程安排 | 文件管理 | 成员权限 | 实时通知",
        "title": "MS Project 任务协作平台 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "MS Project API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:18000",
    "basePath": "/",
    "paths": {
        "/file/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上传文件到项目，支持图片/文档/视频等类型",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "file"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目加密ID",
                        "name": "projectCode",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回文件信息(code/title/fileType/fileSize/fileUrl)",
                        "schema": {
                            "$ref": "#/definitions/common.Result"
                        }
                    },
                    "400": {
                        "description": "参数错误或文件过大",
                        "schema": {
                            "$ref": "#/definitions/common.Result"
                        }
                    }
                }
            }
        },
...
        "/user/login": {
            "post": {
                "description": "使用账号密码登录系统，返回Token和用户信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录信息(account/password)",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回用户信息和Token",
                        "schema": {
                            "$ref": "#/definitions/common.Result"
                        }
                    },
                    "400": {
                        "description": "账号或密码错误",
                        "schema": {
                            "$ref": "#/definitions/common.Result"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "user.LoginReq": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer Token 认证（登录后获取）",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}